Tarvis build Fix

+++ Makefile
@@ -111,7 +111,8 @@ dist: binary clean ## builds source and wheel package
 
 develop: binary
 	python3 -m venv $(VENVDIR)
-	$(BINDIR)/pip install -r requirements_dev.txt
+	test -f $(BINDIR)/pip  && $(BINDIR)/pip install -r requirements_dev.txt
+	test -f $(BINDIR)/pip  || pip install -r requirements_dev.txt
 	$(BINDIR)/python setup.py develop
 
 install: binary clean ## install the package to the active Python's site-packages
